Added Deserialization tests for Locations.
